# docs/swagger.yaml
basePath: /api
definitions:
...
  CreateGuildRequest:
    properties:
      name:
        description: Guild Name. 3 to 30 characters
        type: string
    required:
    - name
    type: object
...
  EditGuildRequest:
    properties:
      icon:
        description: The old guild icon url if no new image is selected. Set null
          the guild icon should be reset
        type: string
      image:
        description: image/png or image/jpeg
        format: binary
        type: string
      name:
        description: Guild Name. 3 to 30 characters
        type: string
    required:
    - name
    type: object
...
  GuildResponse:
    properties:
      createdAt:
        type: string
      default_channel_id:
        type: string
      hasNotification:
        type: boolean
      icon:
        type: string
      id:
        type: string
      name:
        type: string
      ownerId:
        type: string
      updatedAt:
        type: string
    type: object
  JoinRequest:
    properties:
      link:
        type: string
    required:
    - link
    type: object
...
  Member:
    properties:
      color:
        type: string
      createdAt:
        type: string
      id:
        type: string
      image:
        type: string
      isFriend:
        type: boolean
      isOnline:
        type: boolean
      nickname:
        type: string
      updatedAt:
        type: string
      username:
        type: string
    type: object
...
  SuccessResponse:
    properties:
      success:
        description: Only returns true, not an json object
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Valkyrie REST API Specs. This service uses sessions for authentication
  license:
    name: Apache 2.0
  title: Valkyrie API
  version: "1.0"
paths:
...
  /guilds:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/GuildResponse'
            type: array
      summary: Get Current User's Guilds
      tags:
      - Guilds
    post:
      parameters:
      - description: Create Guild
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/CreateGuildRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            items:
              $ref: '#/definitions/GuildResponse'
            type: array
      summary: Create Guild
      tags:
      - Guilds
  /guilds/{guildId}:
    delete:
      parameters:
      - description: Guild ID
        in: path
        name: guildId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/SuccessResponse'
      summary: Leave Guild
      tags:
      - Guilds
    get:
      parameters:
      - description: Guild ID
        in: path
        name: guildId
        required: true
        type: string
      - description: Is Permanent
        in: query
        name: isPermanent
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Get Guild Invite
      tags:
      - Guilds
    put:
      parameters:
      - description: Edit Guild
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/EditGuildRequest'
      - description: Guild ID
        in: path
        name: guildId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/SuccessResponse'
      summary: Edit Guild
      tags:
      - Guilds
...
  /guilds/{guildId}/delete:
    delete:
      parameters:
      - description: Guild ID
        in: path
        name: guildId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/SuccessResponse'
      summary: Delete Guild
      tags:
      - Guilds
  /guilds/{guildId}/invite:
    delete:
      parameters:
      - description: Guild ID
        in: path
        name: guildId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/SuccessResponse'
      summary: Delete all permanent invite links
      tags:
      - Guilds
...
  /guilds/{guildId}/members:
    get:
      parameters:
      - description: Guild ID
        in: path
        name: guildId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/Member'
            type: array
      summary: Get Guild Members
      tags:
      - Guilds
  /guilds/join:
    post:
      parameters:
      - description: Join Guild
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/JoinRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/GuildResponse'
      summary: Join Guild
      tags:
      - Guilds
...
swagger: "2.0"